@@ -1,16 +1,12 @@
-import React from "react/addons";
+import React, { PropTypes } from "react";
 import highlight from "highlight.js";
 import Base from "./base";
 import Radium from "radium";
-import isUndefined from "lodash/lang/isundefined";
 
 @Radium
 class CodePane extends Base {
   createMarkup() {
-    const { source, children } = this.props;
-    // Allow code to come from source or from children, for markdown support
-    const code = (isUndefined(source) || source === "") ? children : source;
-    const markup = highlight.highlight(this.props.lang, code);
+    const markup = highlight.highlight(this.props.lang, this.props.source);
     return {
       __html: markup.value
     };
@@ -21,20 +17,21 @@ class CodePane extends Base {
         <code
           className="hljs"
           style={this.context.styles.components.codePane.code}
-          dangerouslySetInnerHTML={this.createMarkup()}/>
+          dangerouslySetInnerHTML={this.createMarkup()}
+        />
       </pre>
     );
   }
 }
 
 CodePane.contextTypes = {
-  styles: React.PropTypes.object
+  styles: PropTypes.object
 };
 
 CodePane.propTypes = {
-  lang: React.PropTypes.string,
-  source: React.PropTypes.string,
-  style: React.PropTypes.object
+  lang: PropTypes.string,
+  source: PropTypes.string,
+  style: PropTypes.object
 };
 
 CodePane.defaultProps = {
