@@ -1,20 +1,23 @@
-import React from "react/addons";
+import React, { PropTypes } from "react";
 
 const context = function context(Component, params) {
   const wrapper = React.createClass({
     displayName: "ContextWrapper",
     propTypes: {
-      location: React.PropTypes.object,
-      params: React.PropTypes.object
+      location: PropTypes.object,
+      history: PropTypes.object,
+      params: PropTypes.object
     },
     childContextTypes: {
-      styles: React.PropTypes.object,
-      flux: React.PropTypes.object,
-      presenter: React.PropTypes.bool,
-      overview: React.PropTypes.bool,
-      export: React.PropTypes.bool,
-      print: React.PropTypes.bool,
-      slide: React.PropTypes.number
+      styles: PropTypes.object,
+      flux: PropTypes.object,
+      presenter: PropTypes.bool,
+      overview: PropTypes.bool,
+      export: PropTypes.bool,
+      print: PropTypes.bool,
+      slide: PropTypes.number,
+      history: PropTypes.object,
+      location: PropTypes.object
     },
     getChildContext() {
       let styles = {};
@@ -26,10 +29,12 @@ const context = function context(Component, params) {
       }
       let slide = 0;
       if (this.props.params && "slide" in this.props.params) {
-        slide = this.props.params.slide;
+        slide = Number(this.props.params.slide);
       }
       return {
         styles,
+        history: this.props.history,
+        location,
         flux: params.flux,
         presenter: location.query && "presenter" in location.query,
         overview: location.query && "overview" in location.query,
